feat(chat): show degraded-model notice on Safari iOS
The lighter Qwen 0.5B model is already selected for Safari iOS to prevent
OOM crashes. Add a visible amber warning so users know the assistant is
running in reduced-accuracy mode before they enable it and while it is
active.

- Extract `isSafariIOS` to `llm/deviceDetection.ts` so it can be
  statically imported in `useChatEngine` without pulling the WebLLM
  bundle into the main chunk. Re-export it from `engine.ts` to keep
  existing tests and callers unchanged.
- Add `isDegraded: boolean` to the `idle` and `ready` EngineState
  variants. `useChatEngine` sets it once at mount via `isSafariIOS()`.
- `ModelLoader` renders a `DegradedNotice` banner inside the idle card
  (visible before the user enables the assistant) and as a standalone
  notice above the chat when the engine is ready.
